+++ onnxruntime/contrib_ops/cpu/contrib_op_activations.h
@@ -1,0 +1,31 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#pragma once
+
+#include "core/common/common.h"
+#include "core/framework/op_kernel.h"
+#include "core/util/math_cpuonly.h"
+
+namespace onnxruntime {
+namespace contrib {
+
+template <typename T>
+class ScaledTanh final : public OpKernel {
+ public:
+  ScaledTanh(const OpKernelInfo& info)
+      : OpKernel(info), alpha_(info.GetAttrOrDefault("alpha", 1.0f)), beta_(info.GetAttrOrDefault("beta", 1.0f)) {}
+
+  Status Compute(OpKernelContext* context) const override {
+    const Tensor* X = context->Input<Tensor>(0);
+    Tensor* Y = context->Output(0, X->Shape());
+    EIGEN_Y = (T)alpha_ * (EIGEN_X * (T)beta_).tanh();
+    return Status::OK();
+  }
+
+ private:
+  const float alpha_;
+  const float beta_;
+};
+}  // namespace contrib
+}  // namespace onnxruntime
