+++ onnxruntime/contrib_ops/cpu/crop.cc
@@ -1,12 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "core/providers/cpu/tensor/crop.h"
+#include "crop.h"
 
 namespace onnxruntime {
+namespace contrib {
 ONNX_CPU_OPERATOR_KERNEL(
     Crop,
     1,
     KernelDefBuilder().TypeConstraint("T", DataTypeImpl::GetTensorType<float>()),
     Crop<float>);
+}
 }  // namespace onnxruntime
