+++ onnxruntime/contrib_ops/cpu/element_wise_exp_ops.h
@@ -1,0 +1,28 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#pragma once
+
+#include "core/common/common.h"
+#include "core/framework/op_kernel.h"
+#include "core/util/math_cpuonly.h"
+
+namespace onnxruntime {
+namespace contrib {
+template <typename T>
+class Affine final : public OpKernel {
+ public:
+  Affine(const OpKernelInfo& info) : OpKernel(info) {
+    // Either model-supplied or default values should be returned for alpha and beta
+    ORT_ENFORCE(info.GetAttr("alpha", &alpha_).IsOK());
+    ORT_ENFORCE(info.GetAttr("beta", &beta_).IsOK());
+  }
+
+  Status Compute(OpKernelContext* context) const override;
+
+ private:
+  float alpha_;
+  float beta_;
+};
+}  // namespace contrib
+}  // namespace onnxruntime
