+++ onnxruntime/contrib_ops/cpu/image_scaler.cc
@@ -1,12 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "core/providers/cpu/tensor/image_scaler.h"
+#include "image_scaler.h"
 
 namespace onnxruntime {
+namespace contrib {
 ONNX_CPU_OPERATOR_KERNEL(
     ImageScaler,
     1,
     KernelDefBuilder().TypeConstraint("T", DataTypeImpl::GetTensorType<float>()),
     ImageScaler<float>);
+}
 }  // namespace onnxruntime
