+++ onnxruntime/contrib_ops/cpu/image_scaler.h
@@ -8,6 +8,7 @@
 #include "core/util/math_cpuonly.h"
 
 namespace onnxruntime {
+namespace contrib{
 
 template <typename T>
 class ImageScaler final : public OpKernel {
@@ -51,5 +52,5 @@ class ImageScaler final : public OpKernel {
   float scale_;
   std::vector<float> bias_;
 };
-
+}
 }  //namespace onnxruntime
