+++ onnxruntime/core/framework/environment.cc
@@ -40,38 +40,6 @@ Status Environment::Initialize() {
       RegisterOnnxOperatorSetSchema();
       RegisterOnnxMLOperatorSetSchema();
     });
-    //TODO:put all of the following things into call_once
-    // Register MVN operator for backward compatibility.
-    // Experimental operator does not have history kept in ONNX. Unfortunately, RS5 takes bunch of experimental operators
-    // in onnx as production ops. MVN is one of them. Now (9/26/2018) MVN is a production function in ONNX. The experimental
-    // MVN op was removed. The history has to be kept locally as below.
-    ORT_ATTRIBUTE_UNUSED ONNX_OPERATOR_SCHEMA(MeanVarianceNormalization)
-        .SetDoc(R"DOC(Perform mean variance normalization.)DOC")
-        .Attr("across_channels", "If 1, mean and variance are computed across channels. Default is 0.", AttributeProto::INT, static_cast<int64_t>(0))
-        .Attr("normalize_variance", "If 0, normalize the mean only.  Default is 1.", AttributeProto::INT, static_cast<int64_t>(1))
-        .Input(0, "input", "Input tensor of shape [N,C,H,W]", "T")
-        .Output(0, "output", "Result, has same shape and type as input", "T")
-        .TypeConstraint(
-            "T",
-            {"tensor(float16)", "tensor(float)", "tensor(double)"},
-            "Constrain input and output types to float tensors.")
-        .TypeAndShapeInferenceFunction(propagateShapeAndTypeFromFirstInput);
-
-    ORT_ATTRIBUTE_UNUSED ONNX_OPERATOR_SCHEMA(ScaledTanh)
-        .Attr("alpha", "Scaling value", AttributeProto::FLOAT, OPTIONAL)
-        .Attr("beta", "Scaling value", AttributeProto::FLOAT, OPTIONAL)
-        .Input(0, "input", "Input tensor", "T")
-        .Output(
-            0,
-            "output",
-            "The scaled hyperbolic tangent values of the input tensor "
-            "computed element-wise",
-            "T")
-        .TypeConstraint(
-            "T",
-            {"tensor(float16)", "tensor(float)", "tensor(double)"},
-            "Constrain input and output types to float tensors.")
-        .TypeAndShapeInferenceFunction(propagateShapeAndTypeFromFirstInput);
 
     // Register MemCpy schema;
