+++ onnxruntime/core/providers/cpu/activation/activations.cc
@@ -19,16 +19,13 @@ namespace onnxruntime {
 REGISTER_UNARY_ELEMENTWISE_KERNEL(Elu, 6);
 REGISTER_UNARY_ELEMENTWISE_KERNEL(HardSigmoid, 6);
 REGISTER_UNARY_ELEMENTWISE_KERNEL(LeakyRelu, 6);
-REGISTER_UNARY_ELEMENTWISE_KERNEL(ParametricSoftplus, 1);
 REGISTER_UNARY_ELEMENTWISE_KERNEL(Relu, 6);
-REGISTER_UNARY_ELEMENTWISE_KERNEL(ScaledTanh, 1);
 REGISTER_UNARY_ELEMENTWISE_KERNEL(Selu, 6);
 REGISTER_UNARY_ELEMENTWISE_KERNEL(Sigmoid, 6);
 // SoftPlus is the default case for ParametricSoftPlus
 REGISTER_UNARY_ELEMENTWISE_KERNEL_ALIAS(Softplus, ParametricSoftplus, 1);
 REGISTER_UNARY_ELEMENTWISE_KERNEL(Softsign, 1);
 REGISTER_UNARY_ELEMENTWISE_KERNEL(Tanh, 6);
-REGISTER_UNARY_ELEMENTWISE_KERNEL(ThresholdedRelu, 1);
 REGISTER_UNARY_ELEMENTWISE_KERNEL(ThresholdedRelu, 10);
 
 template <>
