+++ onnxruntime/core/providers/cpu/math/element_wise_ops.h
@@ -269,22 +269,6 @@ class Mean_8 final : public OpKernel {
   Status Compute(OpKernelContext* context) const override;
 };
 
-template <typename T>
-class Affine final : public OpKernel {
- public:
-  Affine(const OpKernelInfo& info) : OpKernel(info) {
-    // Either model-supplied or default values should be returned for alpha and beta
-    ORT_ENFORCE(info.GetAttr("alpha", &alpha_).IsOK());
-    ORT_ENFORCE(info.GetAttr("beta", &beta_).IsOK());
-  }
-
-  Status Compute(OpKernelContext* context) const override;
-
- private:
-  float alpha_;
-  float beta_;
-};
-
 // PRelu is activation function, but it's closer to binary elementwise ops in implementation
 template <typename T>
 class PRelu final : public OpKernel {
